\name{bcSweave}
\alias{bcSweave}
\title{Wrapper for Sweave}
\description{Wrapper for Sweave - allows to set options, or perform
  additional preprocessing or QC steps.}
\usage{bcSweave(f)}
\arguments{
  \item{f}{character, filename. Must have extension \code{.Rnw}}
}
\details{
  No details.
}

\value{
  Function is called for its side effects.
}

\author{Wolfgang Huber <huber@ebi.ac.uk>}
\examples{
  ## bcSweave("Intro.Rnw")
}

\keyword{misc}
