\name{perf}
\alias{performance}
\alias{Survperformance}
\alias{ALLperformance}
\alias{kidpackperformance}
\docType{data}
\title{Results of tedious computations for computational inference}
\description{
Results of tedious computations for computational inference
}
\usage{
data(Survperformance)
}
\details{
These datasets are used in conjunction with the
computational inference chapter of the monograph
}
\value{
these are \code{data.frame} objects
}
%\references{  }
\author{Vince Carey <stvjc@channing.harvard.edu>}
%\note{  }
%
%
%\seealso{  }
%
\examples{
data(Survperformance)
objects()
dim(performance)
}
\keyword{data}
